size

+++ 0x00-hello_world/6-size.c
@@ -13,10 +13,10 @@ int main(void)
 	long long int d;
 	float f;
 
-printf("Size of a char: %u byte(s)\n", (unsigned long)sizeof(a));
-printf("Size of a int: %u byte(s)\n", (unsigned long)sizeof(b));
-printf("Size of a long int: %u byte(s)\n", (unsigned long)sizeof(c));
-printf("Size of a long long int: %lu byte(s)\n", (unsigned long)sizeof(d));
-printf("Size of a float: %u byte(s)\n", (unsigned long)sizeof(f));
+printf("Size of a char: %u byte(s)\n", (long unsigned long)sizeof(a));
+printf("Size of a int: %u byte(s)\n", (long unsigned long)sizeof(b));
+printf("Size of a long int: %u byte(s)\n", (long unsigned long)sizeof(c));
+printf("Size of a long long int: %lu byte(s)\n", (long unsigned long)sizeof(d));
+printf("Size of a float: %u byte(s)\n", (long unsigned long)sizeof(f));
 return (0);
 }
